requirements.txt
easydict==1.9
glog==0.3.1
loguru==0.5.3
numpy==1.18.5
matplotlib==3.3.2
opencv-python==4.4.0.44 
PyYAML==5.3.1
scipy==1.5.2
scikit-learn==0.23.2
tensorflow-gpu==1.15.4
tqdm==4.49.0